analyzer/requirements.txt
kafka-python
python-dotenv
reverse-geocode